+++ tests/User/UserServiceTest.php
@@ -45,6 +45,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getTitle());
         $this->assertSame('', $user->getFirstname());
         $this->assertSame('', $user->getLastname());
+        $this->assertSame(null, $user->getBirthday());
 
         // shipping address
         $this->assertSame('', $user->getShippingAddress()->getTitle());
@@ -95,6 +96,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getTitle());
         $this->assertSame('Jean', $user->getFirstname());
         $this->assertSame('Paul', $user->getLastname());
+        $this->assertSame(null, $user->getBirthday());
 
 
         $userService->updateUser(
@@ -104,6 +106,7 @@ final class UserServiceTest extends ApiTestCase
                 ->setFirstName('Jacques')
                 ->setLastName('Jules')
                 ->setTitle(UserTitle::MR())
+                ->setBirthday(\DateTime::createFromFormat('U', '298652400'))
         );
 
         $client->authenticate('user43@example.com', 'password');
@@ -113,6 +116,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertTrue(UserTitle::MR()->equals($user->getTitle()));
         $this->assertSame('Jacques', $user->getFirstname());
         $this->assertSame('Jules', $user->getLastname());
+        $this->assertSame('298652400', $user->getBirthday()->format('U'));
     }
 
     public function testUpdateUserAddresses()
@@ -262,6 +266,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertSame(null, $user->getTitle());
         $this->assertSame('Jean', $user->getFirstname());
         $this->assertSame('Paul', $user->getLastname());
+        $this->assertSame(null, $user->getBirthday());
 
         $userService->updateUser(
             (new UpdateUserCommand())
@@ -278,6 +283,7 @@ final class UserServiceTest extends ApiTestCase
         $this->assertNull($user->getTitle());
         $this->assertSame('Jacques', $user->getFirstname());
         $this->assertSame('Jules', $user->getLastname());
+        $this->assertNull($user->getBirthday());
     }
 
     public function testRecoverPassword()